Drop NFS support
I found Vagrant's integration with NFS to be unreliable. Deciding whether or
not to use it at run-time will entail supporting two rather divergent setups,
each of which is likely to exhibit its own set of problems. Reliability and
portability trumps performance, so I'm switching to exclusively using
VirtualBox Shared Folders by default. It's still possible to make a simple edit
to the Vagrantfile to get NFS, but I won't support it.

Change-Id: I31a24d94f6f207010901831e0fa6547b3a6f5d46