Use ppa to update git, then clone MediaWiki normally
This change tells puppet to add the git-core PPA to apt prior to installing
git. This bumps the version of git installed by puppet to 1.8, which fixes the
problems that made cloning the repository using 1.7.9.5 unfeasable. With Chad's
work on making git-clones of MediaWiki core leaner, it is now sensible to clone
the entire history.

Change-Id: I9742facc682bb487a25760a522252e4d44cc59e9