Make shared data accessible under a single sharepoint
Vagrant 1.1.5 disallows mounting a share (or subtree of a share) under two
different mountpoints. I was using that as a trick to work around flakiness
on Windows when symlinking /var/www/w -> /mediawiki/vagrant. This change
removes both the workaround and the need for a symlink via a couple of
mod_rewrite directives.

The multiple mountpoint thing was also blocking using NFS on Linux / OS X
hosts, so I'll do that next.

Also adds a small notice regarding success/failure of detecting VBox version.

Change-Id: I9aab2fec5bf8147b7cf25df430dab0fcd740d55b