Tell Puppet not to time out when git-cloning core
Sadly Puppet's default 300 second timeout on any Exec[] is not always generous
enough. Rather than set a high timeout, I set it to '0'. The reason being that
someone on a very slow connection might legitimately want to wait for it to
finish, even if takes a very long time. Rather than put an upper bound on the
timeout, this patch set tells Puppet not to enforce a time limit at all.
Because 'setoutput' is set to true, the user will be able to follow what is
happening. If things are totally stuck, the user can still ctrl-c or close the
window and be no worse off.

Also:
- drop redundant '-U' from 'pip install git-review' (puppet is configured
not to run this if git-review is already installed, so there is no case in
which an upgrade is possible or desired.)
- Make favicon depend on apache2 package.

Change-Id: Ib58526ce2a61140adc0ffa7b004f5d62fbfcdf51