Make FORWARDED_PORT a Facter fact
This change makes the port number of the forwarded port (as configured in
Vagrantfile) available as a Facter fact. The change also makes the MediaWiki
and Apache modules use the Facter fact rather than a hard-coded value. This
makes it quite easy to use a different forwarded port for serving HTTP from the
guest.

Bug: 48474
Change-Id: I6cd95cb7bcb79935c006d68d0ecfd8112c428319