Make file uploads work
Using the images/ subdirectory in the MediaWiki codebase is impractical,
because the MediaWiki folder is shared with the guest, and the synchronization
and persistence of file permissions on network mounts is incredibly tricky.
This change makes the MediaWiki Puppet module explicitly configure the upload
directory.

There is one slightly nasty aspect to this change, which is that 'upload_dir'
is parametrized, but its default value is hard-coded  in ./LocalSettings.php.
LocalSettings.php should probably be managed by puppet (and generated using a
template), but I'd like to do that in a separate patch.

Change-Id: I9287b9fbceeaf236ef2179a7465d11ed701a53c8