Set $HOME when git-cloning
As of git 1.8.1.1 and above, if the home directory can't be accessed, it
results in a fatal error. This becomes a problem if we sudo -u vagrant
git-clone, because sudo does not set $HOME and the vagrant user tries to
read/write /root/.config (which fails).

I thought about this for a while and concluded that the very best thing to do
would be to drop the pretense of generality and simply omit the 'user'
parameter from git::clone, electing instead to always git-clone as 'vagrant'.

Bug: 50143
Bug: 50384

Change-Id: If6a066d0f077e0ea37ef0031ae8a901d3e1b3513