Manage configuration fragments in rc.d-like dir
One of the pain-points with MediaWiki-Vagrant is the way it handles
MediaWiki configuration files. settings/README encourages you to add your own
configuration fragments to settings/*, and if you take the advice you end up
with a directory of PHP files, some of which are managed by Puppet and some of
which are not. This is confusing.

An even more severe problem is the fact that disabling a role (i.e., commenting
it back out in site.pp) does nothing. Puppet stops managing the files that are
associated with the role, but they are left in-place. This becomes especially
problematic when the machine is torn down and recreated, because the globbing
and loading of settings/* means that MediaWiki could try and load leftover
configuration files from a previous instance.

This change improves things somewhat. First, it renames settings/ to
settings.d/, which serves as a hint for people that are familiar with rc.d
style config.d directories. Second, it relegates all the settings files that
are managed by Puppet to settings.d/puppet-managed. Third, it configures Puppet
to purge unmanaged files from the directory on each run. This means that the
state of the directory is always in-sync with what is enabled in Puppet.

The Puppet pattern I used is described here:
<http://christian.hofstaedtler.name/blog/2008/11/puppet-managing-directories-recursively.html>.

Change-Id: I136c46a967183239ba3567767118a0f1692a39fd