Load Puppet-managed settings before user-managed ones; update README
Mr. Flaschen observed that the current load order (user settings, then
Puppet-managed settings) means that in the case of conflict, Puppet wins. This
shouldn't be the case, because we can't expect every user to fiddle with Puppet
manifests in order to configure MediaWiki the way they want to. Flipping the
order fixes that.

This patch also updates settings.d/README to reflect the change.

Change-Id: I77df5fdb5ee43f7b1fccef298b0d922ddb95a255