Perform strict and verbose validation of optional 'Roles.yaml' file
This patch ensures that puppet-classifier fails when it encounters errors in
'Roles.yaml' and that it emits useful error messages.  The previous behavior
was to swallow errors and hand over an empty manifest to Puppet. Matt correctly
noted that swallowing errors could lead to the removal of provisioned
configurations and was therefore unsafe.

Change-Id: If1c751625a972e94ab22b1671e36f78d3392b19f