Tee Puppet output to console & /vagrant/logs/puppet.log
I was not able to find documentation for this, but it appears you can specify
multiple '--logdest' args to Puppet to have the log output go to multiple
destinations. I confirmed this experimentally, and also looked at Puppet's code
to confirm that this behavior is by design. It is, as you can see from
<http://git.io/FAkULw>.

Change-Id: I026c690bf10d29aa6b6cc581b98950c013fea92e