Generate Puppet config_version from short SHA1 of HEAD
This patch tells Puppet to generate the configuration version by capturing the
output of puppet/extra/config-version. That script gets the short SHA1 hash of
HEAD for MediaWiki-Vagrant and the current UTC timestamp and joins them with a
period (for example, '1373576256.89684db'). If it is unable to determine the
Git commit of HEAD, the second field is set to 'no-git' (for example,
'1373576256.no-git').

Puppet uses the configuration version to label its log output. By including the
time and SCM info, it should be easier to correlate Puppet logs to specific
changes. For details, see the Puppet manual: <http://docs.puppetlabs.com/
references/latest/configuration.html#configversion>

Change-Id: Ia11d7cbd686078ccbaca524377addbe1645cf754