Move Vagrant helpers to lib/mediawiki_vagrant.rb
This patch moves various Ruby helper functions that were in Vagrantfile to
lib/mediawiki_vagrant.rb. To make it possible to require the file, the
directory containing Vagrantfile is added to $LOAD_PATH. We can't use
'require_relative', since that's Ruby 1.9+, and Vagrant still supports Ruby
1.8. (This means that loading the Vagrantfile-extra file was also done
incorrectly; I fixed it too in this patch).

This patch also moves Puppet logs from logs/ to logs/puppet, and appends
the short SHA1 of the local MediaWiki-Vagrant's HEAD to the log file name.

Change-Id: I42ea829272a1379e95952d970bf211c2e43e5087