For clarity, use PyYAML to generate output of puppet-classifier
I removed the dependency on PyYAML in I2f805dbded because I thought it was
better to generate the output format manually than depend on a nonstandard
module. But the 'yaml' module is available on Ubunutu by default, and not using
it makes the code less clear, since it isn't obvious why the output is
formatted in the way that it is.

Change-Id: Ic4244dfe97da74cf095702ab1f17a36103600a52