Enable 'box_download_insecure' if Vagrant 1.2.6+
Vagrant 1.2.6 added a 'box_download_insecure' option to allow the Box to be
downloaded via HTTPS despite the fact that Vagrant does not verify SSL certs.
This is the desired behavior from my perspective. The VM is executed in a
sandbox environment by VirtualBox and is expressly not designed to contain
secure data. There is a big, scary warning in the Vagrantfile about it for
paranoid folk.

See <https://github.com/mitchellh/vagrant/commit/af66f5f> for more info.

Change-Id: I3dbc7eb5bc27a4ff85a810f39273470a7cc8043d