Add 'mysql::sql', 'mysql::user' & 'mysql::db' types
This patch adds three custom defines for working with MySQL. The base
define is 'mysql::sql', which is a wrapper around an Exec resource of
the 'mysql' command-line tool that executes arbitrary SQL code.
'mysql::user' and 'mysql::db' provide some additional syntactic sugar by
allowing for the concise creation of databases and users.

Change-Id: I2db984034d541fd31c558e81abe6b6b5fa0908db