Configure UploadWizard for optional chunked upload.
Add configuration to role:uploadwizard that enables the per-user preference
for using chunked uploads. Also adds a few configuration directives shown on
[[Extension:UploadWizard]].

This patch also introduces the ability to specify mediawiki::settings that
need to be expressed as associative arrays. The current implementation should
support values that are numeric, character or boolean literals and/or indexed
arrays of the same. It will not support inner associative arrays at this time.

Change-Id: I5a2a2cd5e45174344159c6b6bf308cd541d8dc32