Add labsvagrant.rb helper script
This is to have a helper script that can run the vagrant
"enable-role", "disable-role", "list-roles", "reset-roles",
"provision" commands on the host machine itself.

Primarily added to help with the labsvagrant module in the
operations/puppet.git repository, to help people use the
puppet modules available in vagrant to setup mediawiki
instances on labs.

Change-Id: I5d6d24d3848b02b2b96cd10e383a7844c5d431f6