Make apache::site for main vhost depend on a2enmod.
The main apache vhost configuration needs some apache modules to be
enabled. This small change tells puppet about this dependency so that it
will run the a2enmod commands prior to installing the vhost
configuration. Without this change order is non-deterministic and the
initial `vagrant provision` for a new instance may fail.

Change-Id: I76428ccf3e3fb4f34fcbd52d4484818b83e14270