Don't point Apache's ErrorLog at vboxsf mountpoint
Brion pointed out that this causes Apache to fail to start when it initializes
before /vagrant is mounted. This could be fixed in other ways that would allow
us to keep the error log accessible, but the bug is severe so I think we'll
have to go with this for now.

Change-Id: Idf4cabbe42cea1a6ac6af35c3fec7450bdf08415