Fix logging with a separate dir and www-data as user and group
The PHP code runs as user and group www-data.  However, neither the
www-data user nor group has write access to /vagrant.

It does need to write to this directory in general, but it does need
to write logs, so this setups a synced directory for that.  The paths
are the same as before in both the host and guest.

Change-Id: I0352239e774a2d17237301b283245b26dddbae02