Add role::profiler utilizing XHProf
Adds a profiling role which installs xhprof from pecl and enables its web
interface on /xhprof of the devwiki http server.

One issue remaining in regards to utilizing disable-role.  The only part that
gets uninstalled when removing the role is the apache configuration.  The php
config installed by php::ini will still be in /etc/php5/conf.d/xhprof.ini and
loaded. The module installed in /usr/lib/php/20090626/xhprof.so will also still
be in place.  Additionally the assets and code files installed to /usr/share/php
are left behind.  In general this isn't a big deal, but worth mentioning.

Change-Id: I12b7041739ba1a4ba9b3c651c99c801123252e56