redis: 16M -> 64M, volatile-lru -> allkeys-lru
As reported by Matthew Flaschen, the trifecta of enabling redis persistence,
leaving the maxmemory-policy to default to volatile-lru, and not setting
volatile keys means that redis will get clogged with data and eventually
decline to set keys. This problem is made especially severe by the fact that
the redis maxmemory setting was set to such a low value (16M).

This patch increases the maxmemory setting to 64M, quadruple the previous value.
It also sets the redis module's default maxmemory-policy to 'allkeys-lru',
since I think that that is the more sensible default.

Bug: 55981
Change-Id: Ia94b5f074cc437ccbcac9e2b9640388e7a2a791e