Add stub HHVM role
I was not clever enough to work out a working configuration for running
MediaWiki on HHVM + Apache + FastCGI, so for now this merely provisions the
package. I'd ordinarily wait to have something more substantial before
committing, but because there are other people interested in working on this, a
stub role has some value.

Change-Id: Ie75c2855326ae0a20a142279e9d1c415351cba3c