Add misc::evergreen resource to refresh services on config updates
If a role that writes its configuration to a recursively-managed foo.d/-type
directory is disabled, its configuration file will be purged, but the service
it configures will not necessarily be refreshed. This patch adds a
misc::evergreen resource that takes a service name and one or more paths to
search. If the most recent modification time of the path or any of the files or
directories within it is newer than the oldest instance of the service, the
service is automatically refreshed.

Change-Id: Ib969eb352461ab655d1b0f368c91e6737122f1c3