Add apache::env
The shell environment defined by /etc/apache2/envvars defines the shell
environment in which the Apache process is spawned. It can be used to define
runtime parameters that can be referenced in Apache configuration files via
<IfDefine> directives. This patch adds a call to /etc/apache2/envvars to source
all scripts in /etc/apache2/env.d, which is recursively managed by Puppet.
A custom resource, apache::env, can be used to plop shell fragments in that
directory.

Change-Id: I3b387fe6c1a7af926f0c8a91b11e6634feb3a8b6