Remove reference to provider.driver.version and virtualbox class
The provider.driver.version message is not supported by the vmware
provider. We were using it to enable checking for outdated VirtualBox
Guest Additions. This is no longer needed since we started bundling the
the vagrant-vbguest plugin in Ib2efe79. Since the virtualbox class only
provided utilities for checking and updating Guest Additions it will be
removed entirely.

Bug: 59904
Change-Id: Ia6bc4f2e53eaa8110509a4053f4ac51a650dfada