Stop 'stdin: is not a tty' complaints
Change the shell command used by vagrant when communicating with the VM
from the default `bash -l` to `bash -c 'BASH_ENV=/etc/profile exec
bash'`. This bypasses the /root/.profile script containing the unguarded
`mesg -n` command which triggers the error message.

This work around is slightly superior to the prior solution of editing
the mesg invocation out of /root/.profile as it takes effect prior to
the initial puppet run.

Change-Id: I635882877e8bbf11599d687e79f82d80a729d213