Update HHVM setup for HHVM 3
Taking Wctaiwan's instructions from
<https://www.mediawiki.org/wiki/HHVM/Vagrant> as a basis, fix the HHVM role so
that it works with the latest packages from Facebook.

It's a bit gross to have so much HHVM-specific things in the MediaWiki module's
Apache site template. I'd like to follow-up with another patch that makes the
HHVM config a matter of "Include hhvm.conf", where hhvm.conf is provided by the
HHVM module.

Change-Id: Iabe31fc3b09501cde179f26f42976d4a8e63f9f2