Add Vagrant plugin to perform reload action
When the Vagrantfile receives changes such as additional port forwards
or changes to memory settings `vagrant reload` must be called to change
the settings of the Virtualbox virtual machine. Recently I689f7c1 added
tooling to make managing such changes with Puppet easier, but stopped
short of automating the application changes. This patch introduces
a Vagrant provisioner plugin that can be signaled to reload the virtual
machine and updates the vagrant::settings Puppet class to raise the
signal when Puppet managed configuration changes.

Bug: 64114
Change-Id: I30a1c5d4702e9d166d8bee3eb0fc843b319ae799