Add a .ackrc file for new vagrant installs, with some new types
ack has the annoying habit that it only searches known file types
unless one specifies --all (this will be fixed in ack 2), and it
lacks knowledge of some important types like LESS. This leads to
easy mistakes when refactoring. This commit adds the most
problematic missing types: .less for LESS, .json for the
localization files, .pp for puppet. It does not touch existing
.ackrc files since there is no clean way of merging contents.

Change-Id: Icaf783c74cea9f23dfc6b0f75d8664cae4c529ef