Support for multiple wikis
The multiwiki class and related defines provide support for creating and
managing a small wiki farm on a MediaWiki-Vagrant instance. Each wiki
will share the same MediaWiki and extension code but have independent
LocalSettings.php configuration and databases. These features can be
used to create new roles for MediaWiki-Vagrant that provide complex
features such as CentralAuth and WikiData.

See Ide7acba for a usage example that enables CentralAuth on the main
MediaWiki-Vagrant wiki and adds two additional wikis.

This feature was initially developed during the Zürich hackathon and
benefited from the help and experience of many people in attendance.

Co-author: csteipp <csteipp@wikimedia.org>
Change-Id: Id10c890da73a022649b2f493bf49d75e1af0eb35