Revert "Set proper SSH URL for all GIT repos (with username)"
This reverts commits c97683592a0 and 08c9a2c840b.

This ended up causing a lot of pain for users and it's still quite broken. It
violates the principle that we previously adhered to of not taking liberties
with the user's git setup, especially since the consequences of that spilled
onto the host.

I'd like to remove this for now until we come up with a better solution.

Change-Id: Iaf1247dd6440908b977001b8c465a0cc67245be2