Packaged plugin as gem to allow local installation
Since Vagrant 1.6, plugins can not be loaded from within the
Vagrantfile. Packaging the plugin as a gem provides a workaround for
users of 1.6 and up.

In addition to providing a workaround, having at least a gemspec to
reference in the Gemfile fits better with the development workflow
suggested in the Vagrant docs.

Removed direct use of plugin features within Vagrantfile to allow for
basic usage without requiring plugin installation. Edited README to
communicate that the plugin is optional.

Bug: 65066
Change-Id: Id42866a80a650f87319f445aafc897ada3cb618b