Fix roles_available parsing
The `vagrant list-roles` output and other plugin based role management
actions are driven in part by the roles_available method in the
helpers.rb file. The addition of parameters to role::browsertests
revealed that the string extraction done by roles_available was overly
greedy in its matching. This led to the browsertests role being listed
as "browsertests(" in the list-roles output.

This change corrects the parser by changing from non-whitespace matching
to word-character matching. It also adds whitespace to the role
declaration for good measure. The version number of the
mediawiki-vagrant plugin is also bumped to indicate that the
functionality of the plugin has changed.

Change-Id: I96979a0c93f8b92b97919f024dbd21d74680159f