Fixes to Apache module
I thought we could do away with the recursively-managed directory of Apache
environment scripts, but the issue with the apache::def resource is that
disabling a role that declares it leaves the variable declared. So, for the
moment, you can't really disable the zend role once you enable it.

To fix it, bring back apache::env, and make apache::def delegate to it.
Because people using the current Puppetization may already have variables
defined in /etc/apache2/envvars, we need to reset it. So just bite the bullet
and provision our own copy (which is basically the default envvars file, with
the comments stripped, and the loop to iterate through env-enabled, which we
were previously adding via a sed/grep exec).

Also remove NameVirtualHost directives (they have no effect under Apache 2.4
other than to trigger a warning).

Bug: 67683
Change-Id: Id4b56940f634081c01d4fbabca14dd65f2b85c06