hhvm: set GzipCompressionLevel = 0 to avoid double-gzip
MediaWiki does its own gzipping. Having HHVM do it as well generates responses
that are twice-compressed and sent with the header "Content-Encoding: gzip, gzip".
In a more perfect universe, gzipping something twice would compress it twice as
well. In this universe, it inflates the response size and breaks content
decoding on Safari.

Bug: 67928
Change-Id: Ibc703ac4c2b3852c1b9afc49ac8f74a4df620247