Enforce order of Apt changing files without comparison
Apparently Puppet does not consider Packages being installed using the
default provider as matches for a collection using a comparison based on
the provider. Seen when using "file {} -> Package <| provider == apt |>":

  ==> default: Notice:
    /Stage[main]/Role::Math/Package[texlive-lang-german]/ensure
    : ensure changed 'purged' to 'present'
  ==> default: Notice:
    /Stage[main]/Packages::Php5_tidy/Package[php5-tidy]/ensure:
     ensure changed 'purged' to 'present'
   ==> default: Notice:
     /Stage[main]/Main/File[/etc/apt/apt.conf.d/01no-recommended
     ]/ensure: defined content as
     '{md5}bcf7b831e0adb38a371ce5662a65104f'

When the collection is changed to omit the comparison and apply to all
Package resources Puppet enforces the desired ordering:

  ==> default: Notice:
  /Stage[main]/Main/File[/etc/apt/apt.conf.d/01no-recommended]/ensure:
  defined content as '{md5}53ce69211afabdf3f0123353515e80f4'
  ==> default: Notice:
  /Stage[main]/Role::Math/Package[texlive-latex-extra]/ensure: ensure
  changed 'purged' to 'present'
  ==> default: Notice:
  /Stage[main]/Role::Math/Package[texlive-publishers]/ensure: ensure
  changed 'purged' to 'present'

This behavior is difficult to spot in testing as it only affects the
first Puppet run on a VM after the Apt configuration resources are
introduced.

Change-Id: I47d9e7f80cfddd69f48fe4d16b3e9f509ac2680a