Update xhprof module
This module was constantly reinstalling itself due to the update
to trusty.  This was due to the change in binary api number.
xhprof now looks for the appropriatly generated binary, although
its still a fragile check.

The Apache::conf method has changed, it no longer has a site
parameter. The current default operation of apache::conf does what
we want, add global configuration, so no other change is necessary.

Change-Id: I4b814943bfc9144667c6a7eea3e60d451bdbb610