Enable runtime fcgi debugging
In order for the cli debugger to attach to web requests
the hhvm.debug.enable_debugger parameter needs to be enabled.  Additionally
for the debugger to be able to attach to our current web requests hhvm needs
a small patch[1] which has been merged and will be in the next release.

With this enabled the process is:

    $ hhvm -m debug --debug-host localhost --debug-port 8089
    localhost> machine list
    machine list
      1     vagrant's default sandbox at /vagrant/www/
      2     __builtin's default sandbox at /
    localhost> machine attach 2
    machine attach 2
    Attaching to __builtin's default sandbox at / and pre-loading, please wait...
    localhost> break MediaWiki::run()
    break MediaWiki::run()
    Breakpoint 1 set upon entering MediaWiki::run()
    localhost> continue
    continue

At this point the next web request that enters MediaWiki::run() (all of them)
will break and put you on a command prompt:

    Breakpoint 1 reached at MediaWiki::run() on line 450 of /vagrant/mediawiki/includes/MediaWiki.php
     449            try {
     450*                   $this->checkMaxLag();
     451                    try {
    localhost>

[1] https://github.com/facebook/hhvm/commit/9561315d8b8eafc8f5c3b976b7fc027495409545

Change-Id: I3d60d787259d95b946fdda7c757cb8859e9d20fe