Set up port forward for Wikimetrics automatically
Wikimetrics required manual mangling of Vagrantfile to set up the
forward for port 5000. This was tedious, and meant that git always
thought that Vagrantfile is modified. Hence, we now set up the forward
withing the wikimetrics role directly.

Thereby, Vagrantfile stays clean, and wikimetrics works out of the
box. No additional manual setup is required after enabling and
provisioning the role.

Change-Id: I952d3cdcbae2c34cfd80b83af7ef8cd861c5fb6f