Set $wgServer from inbound request
Auto-detecting $wgServer via WebRequest::detectServer() is not optimal
in a production server, but it can be very useful in a development
environment. Taking the current values from the Apache container rather
than canonically redirecting traffic to http://127.0.0.1:8080/ allows
easier cross-vm communications and the possibility to share the primary
wiki via `vagrant share`.

Bug: 66399
Change-Id: I8337b4a110d52b4fe3c252af738dd0394f3f851f