Use the largest of all configured vagrant_ram values
Modified the coercion lambda for vagrant_ram to take the larger of the
new value and the currently configured one (as opposed to just the
new or the default).

Fixed config command to unset the old value before parsing the new so
that local values for settings such as vagrant_ram can be lowered by
re-running `vagrant config`.

Change-Id: I280f85ba70bf818b4a917e902d32b574b200f287
Bug: 69720