Provision mobile wiki with role::mobilefrontend
Add an always mobile skinned wiki when role::mobilefrontend is enabled.
The existing behavior of user-agent based theme switching is preserved
for other wikis including the default wiki. Switching on the new
"mobile.wiki.local.wmftest.net" is managed by setting $wgMFMobileHeader
to "Host" which will always be present in the request.

Change-Id: Icf55f679d1592aa68e04d294510f7d0f612791bc