Uploading needs login to work, so configure that
Some configurations of mediawiki allow anyone to upload
files, but in the current mediawiki-vagrant configuration, that
won't work - the file repo demands all files must be owned.

(I've had some strange errors occur in automated testing because
the user didn't log in first or login failed, but was still
allowed to upload.)

So, we add the configuration requiring users to log in
to upload.

All tests assumed that users were logged in anyway. And
this allows us to test the error messages that occur when
users don't have appropriate permissions.

Change-Id: I6bcfaac69df3abb544f3c79db40330ff375c9e0b