Revert "Set _SERVER['REMOTE_ADDR'] in cli mode"
This reverts commit 7058f989c5d57d44701d3f4d4487e792a7a9680e.

This is incorrect and breaks at least one unit test. IP addresses
shouldn't be needed in command line mode. If something is failing
because of an IP then that code path should have its assumptions
checked.

I'm not even sure that the original code for GetIP is needed--
what's the bug we're trying to avoid here? Why don't we have similar
problems in production?

Change-Id: I288974f6cb8899862eccd6dba5a37adba101855e