Wiki installer now conditional on existence of database
Removed application-level check and subsequent removal of
LocalSettings.php in favor of an 'unless' condition that includes
checking for the existence of the wiki database which the setup script
is assumed to create. The former is too fragile and can be falsely
triggered by user-level factors (e.g. botched merges or rebases).

Bug: 50912
Bug: 70314
Change-Id: Ia2ff2ab53966ade7bbacd705dc43369345d1b0e0