Split HHVM php.ini into common and fcgi hiera settings
Refactored hhvm class to write separate common and fcgi-specific php.ini
files based on settings pulled in from hiera. This approach is similar
to that of the ops version but use of hiera over puppet hashes should
allow for greater flexibility in configuration and a more generalized
manifest.

The max_execution_time setting was moved to the FCGI settings to avoid
execution timeouts in CLI scripts (e.g. composer).

Change-Id: Ifbfc38200808053d1b9700b9260d50110a5eccfa
Bug: 70843