Disable puppet service
The puppet package for Ubuntu 14.04 has changed behavior to ensure that
puppet agent is started by default. Since we are masterless and invoke
puppet manually via `vagrant provision` there is no need to have this
behavior in the VM. We do have an interesting problem however in that we
should not disable the puppet agent when invoked via labs-vagrant. In
that environment running a puppet agent is very desirable as puppet is
used to manage content outside of the purview of our manifests.

Bug: 70421
Change-Id: I69db5b1c9af3d3e1db66570337075e9d54850b29