Configure wiki cache and upload directories
Added provisioning of cache and upload directories to mediawiki::wiki
and modified wgConf template to include the settings. The upload
directory will continue to be shared between wikis (migration of
existing files is problematic) but the cache directory now defaults to a
subdirectory of $::mediawiki::cache_dir.

Removed $wgCacheDirectory from LocalSettings.php to avoid confusion.

Bug: 70959
Change-Id: I45def381a30c5c0719ab38ca627c7c91df91c5e3