Sync /etc/apt/sources.list.d/wikimedia.list with WMF production
On a labs_vagrant instance, the wikimedia.list installed by
operations/puppet and ours disagree. This causes unnecessary battles
between the two puppet systems for the "proper" content of this file.

Add non-free component and remove comments about puppet managing the
file so that our version of the file looks like theirs.

Change-Id: I38f4d91dcefc8c48c34c492f2d23d0dd15474b36