Enable XDebug with HHVM
This works for me, but not reliably.  For example, I get
https://github.com/facebook/hhvm/issues/3926 and not all debugging
commands seem to respond correctly.

However, it should not interfere with the experience if you are not
using XDebug (i.e. if the browser cookie is not set).

Thus, I think it's mergable so people can continue to attempt to use
it and troubleshoot the debugger itself.

Bug: 71777
Change-Id: I2e5818012e7bd0d52be8a1eb2a02b96bdb341c6d