hhvm: Cleanup configuration
The use of hdf files is no longer required. The operations/puppet
classes have been fixed for this for a while, so it was used as
a template update our class. Other small changes were made to bring our
hhvm class closer to the one that is used in operations/puppet.

A few additional HHVM ini settings were added that may help with local
debugging. Then again they may not. Reading HHVM source code to discover
what settings are used for is a bit of a dark art that I haven't quite
mastered yet.

Change-Id: Ic5b27158c322c0f20844f5c05d140ec77eb86e51