Prefer source for extension's composer install
This adds an option in the composer install class to
specify preferred source. ('dist' or 'source')

For extensions (at least), it makes sense to use source
since vagrant is often used for development purposes.
This makes it easier to contribute patches to these
libraries or extensions (e.g. Wikibase) that are installed
this way.

Change-Id: I6d5aad804e28dd1eff54c7258ce90914549d1229