plugin: Optional display of only enabled roles
Add new options to `vagrant roles list`:
* `--enabled`: only show enabled roles
* `--onecol`: show output in a single column without headers/footers
* `--quiet`: suppress header and footer information

Among other things this enables the `vagrant roles list -e -1`
variation that will list all enabled roles in a single column of output
with no report header or footer. This output is suitable for use in
a shell pipeline that wants to process a list of enabled roles.

The command assumes `--quiet` and `--onecol` as the default for output
to non-interactive terminals (i.e. piped or redirected).

Change-Id: I19160e9339cd8e10000df5b0ea9e6b544c7b9263