plugin: Support "auto" setting for vagrant_ram and vagrant_cores
Giving "auto" as the value for vagrant_ram will now automatically set
the value to 1/4th of total system memory. Note that the value is only
calculated once to avoid repeatedly calling out to shell commands
(especially slow for the Windows implementation).

Setting vagrant_cores to "auto" will make all of the host computer's
cores available to the virtual machine.

Bug: T87684
Change-Id: Iddf48526b2128d7e8ba990405b2c1f39c3ce1956