Make sure all PHP files are available from the host machine
Copy PHP files from the docroot to a subdirectory of /vagrant.
(Simply placing the docroot there would run into file permission
problems as vagrant does not have root access to NFS shares.)

This is important for debuggers which can get confused by remote
files not present on the host machine.

Bug: T88149
Change-Id: I7058f60017226ad2a73791bee7fa1b37013630a9