Load extensions and skins using extension registration if possible
If the extension or skin has a extension.json/skin.json file, and
the 'ExtensionRegistry' class exists, those extensions will be loaded
through the extension registration system rather than the PHP entry point.

The requirement for SkinFactory to be present when loading skins was removed
since Vector no longer depends upon it.

Bug: T86990
Change-Id: I8175d159017c41db6e52bfff63c66c3fe70e2448